announce 3040/3060 support, fully fix mac addresses for xx60/xx40
the code for the mac address configuration in 1e6334dbfae335088039d89f57c7f72c0fa5850e was correct, however it was in the wrong place. it should have been in embedded/common/eeprom.c and the destination should be pAd->PermanentAddress

i made this fix, tested it on the 3040/3060, whose SES button functionality and LEDS have been tested and working.

I will hang on to the 3040 for a few days to see if anyone cares to give feedback. 

just remember one thing: the DBDC mode will limit the number of VAPS. you can try to add 15 additional VAPs for ra0 and rax0 separately, but you will only get 15 total. same goes for the WDS interfaces. this is because it is one radio with two bands, not an entirely separate radio.

link speeds have been tested via homedale and are reported at slightly higher than AC3000. 

builds are coming shortly.