update to 47146, fix 47111/47143 httpd status bug
in 47111/47143, it turns out i removed a line that i thought wasn't
necessary in the services init code.

this line set wl0_ifname to ra0, and i didn't see why this was needed
because wl1_ifname gets set without it.

while the mystery as to how this happens continued, two users discovered
that restoring from defaults would cause the httpd status page to hang,
confirming that wl0_ifname was unset and thus failing to parse the
temperature for the radio (and hanging the status page).

it has now been fixed. thank you nikolatesluh and markonius