small fix for eeprom file locations
this will probably not have any measurable effect, but it is possible.
just fixing the location of the name of some eeproms in case it helps.

i personally have gotten the best speeds ever on the build uploaded
previously. it's possible this one will help but at worst it's just
lateral.