small README fix + additional note on 47182
just wanted to note that i reverted the ramcode on 47182 to what
@hanwckf is using. i noticed that it would randomly drop the download
speed when you're connected for a while. so just wnated to make mention
of that.

"Sure we'd be happy to take your money. Yup, just go to our Kickstarter
page... Okay nice idiot, Uh-huh fuck you, bye bye."

and "from one redskin to another, go fuck yourself!"