[4.14.313/52382]🔥🔥fix last bit of httpd "connected clients" bug🔥🔥
noticed that some of the (V)APs weren't showing the right user count,
and noticed there was a bug.

forgot to remove one line of code to ensure all of the connected clients
are displayed correctly for each (virtual) AP.

i don't think there's a way to do packet stats for each VAP at this
time, but don't hold me to it.

see previous commit for full changelog. a bunch of shit that i didn't
mention was updated too (openvpn among others)