[v57595] update smartdns, openvpn, whatever else
here is a monthly update. it should work. if there are any distortions
on the info screen with the temp sensors, let me know and i'll try to
fix it.

one day i'll try moving all builds to 6.x but im a little busy right now
with some regulatory stuff and also i have to do some testing to get
firefox working on 10.7, so this is third on the list.

later