[v59421] fix upnp issue.
this one shouldn't break 2640/3040 if it does, open a ticket
and i'll fix it. sometimes assfuck hacks the shit out of the nand stuff
in the linux kernel since he loves to support PIECE OF SHIT
architectures with PIECE OF SHIT designs that are best shown by their
PIECE OF SHIT dts files.

upnp wasn't showing up in the gui before but now it's fine. my bad